+++ src/main/java/com/avairebot/commands/music/PauseCommand.java
@@ -81,6 +81,10 @@ public class PauseCommand extends Command {
             return sendErrorMessage(context, context.i18n("error", generateCommandPrefix(context.getMessage())));
         }
 
+        if (!musicManager.canPreformSpecialAction(this, context, "pause music")) {
+            return false;
+        }
+
         if (musicManager.getPlayer().isPaused()) {
             context.makeWarning(context.i18n("alreadyPaused"))
                 .queue(message -> message.delete().queueAfter(1, TimeUnit.MINUTES, null, RestActionUtil.ignore));
